+++ changelog.md
@@ -4,6 +4,7 @@
 - Add jam field for selecting forms on entities
 - Add radio field type
 - Separate form partial for manual, non view composer usage
+- Change save and exit to save and create fields on forms with no fields
 
 ## Version 2.3.6 (2017-02-08)
 - Fix validation for files